+++ libraries/microsoft-agents-hosting-teams/microsoft_agents/hosting/teams/teams_agent_extension.py
@@ -28,720 +28,6 @@ from microsoft_teams.api.models import (
     TaskModuleRequest,
 )
 
-StateT = TypeVar("StateT", bound=TurnState)
-
-CommandSelector = str | Pattern[str] | None
-
-
-def _match_selector(selector: CommandSelector, value: Optional[str]) -> bool:
-    if selector is None:
-        return True
-    if value is None:
-        return False
-    if isinstance(selector, str):
-        return selector == value
-    return bool(re.match(selector, value))
-
-
-def _get_channel_event_type(context: TurnContext) -> Optional[str]:
-    data = context.activity.channel_data
-    if data is None:
-        return None
-    if isinstance(data, dict):
-        return data.get("eventType") or data.get("event_type")
-    return getattr(data, "event_type", None)
-
-
-async def _send_invoke_response(context: TurnContext, body: Any = None) -> None:
-    serialized_body = None
-    if body is not None:
-        if hasattr(body, "model_dump"):
-            serialized_body = body.model_dump(
-                mode="json", by_alias=True, exclude_none=True
-            )
-        else:
-            serialized_body = body
-    await context.send_activity(
-        Activity(
-            type=ActivityTypes.invoke_response,
-            value=InvokeResponse(status=int(HTTPStatus.OK), body=serialized_body),
-        )
-    )
-
-
-class MessageExtension(Generic[StateT]):
-    """
-    Route registration for Teams Message Extension (composeExtension) invoke activities.
-    Access via TeamsAgentExtension.message_extension.
-    """
-
-    def __init__(self, app: AgentApplication[StateT]) -> None:
-        self._app = app
-
-    def on_query(
-        self,
-        command_id: CommandSelector = None,
-        *,
-        auth_handlers: Optional[list[str]] = None,
-        rank: RouteRank = RouteRank.DEFAULT,
-    ) -> Callable:
-        """Register a handler for composeExtension/query invokes."""
-
-        def __selector(context: TurnContext) -> bool:
-            if (
-                context.activity.type != ActivityTypes.invoke
-                or context.activity.name != "composeExtension/query"
-            ):
-                return False
-
-            value = context.activity.value
-            command_value: Optional[str] = None
-            if isinstance(value, dict):
-                command_value = value.get("commandId") or value.get("command_id")
-            elif value is not None:
-                command_value = getattr(value, "commandId", None) or getattr(
-                    value, "command_id", None
-                )
-
-            return _match_selector(command_id, command_value)
-
-        def __call(func: Callable) -> Callable:
-            async def __handler(context: TurnContext, state: StateT) -> None:
-                query = MessagingExtensionQuery.model_validate(
-                    context.activity.value or {}
-                )
-                response = await func(context, state, query)
-                if response is not None:
-                    await _send_invoke_response(context, response)
-
-            self._app.add_route(
-                __selector,
-                __handler,
-                is_invoke=True,
-                rank=rank,
-                auth_handlers=auth_handlers,
-            )
-            return func
-
-        return __call
-
-    def on_select_item(
-        self,
-        handler: Optional[Callable] = None,
-        *,
-        auth_handlers: Optional[list[str]] = None,
-        rank: RouteRank = RouteRank.DEFAULT,
-    ) -> Callable:
-        """Register a handler for composeExtension/selectItem invokes."""
-
-        def __selector(context: TurnContext) -> bool:
-            return (
-                context.activity.type == ActivityTypes.invoke
-                and context.activity.name == "composeExtension/selectItem"
-            )
-
-        def __register(func: Callable) -> Callable:
-            async def __handler(context: TurnContext, state: StateT) -> None:
-                response = await func(context, state, context.activity.value)
-                if response is not None:
-                    await _send_invoke_response(context, response)
-
-            self._app.add_route(
-                __selector,
-                __handler,
-                is_invoke=True,
-                rank=rank,
-                auth_handlers=auth_handlers,
-            )
-            return func
-
-        if handler is not None:
-            return __register(handler)
-        return __register
-
-    def on_submit_action(
-        self,
-        command_id: CommandSelector = None,
-        *,
-        auth_handlers: Optional[list[str]] = None,
-        rank: RouteRank = RouteRank.DEFAULT,
-    ) -> Callable:
-        """Register a handler for composeExtension/submitAction invokes (not bot message preview)."""
-
-        def __selector(context: TurnContext) -> bool:
-            if (
-                context.activity.type != ActivityTypes.invoke
-                or context.activity.name != "composeExtension/submitAction"
-            ):
-                return False
-            value = context.activity.value
-            if isinstance(value, dict):
-                bot_message_preview_action = value.get("botMessagePreviewAction")
-                resolved_command_id = value.get("commandId") or value.get("command_id")
-            else:
-                bot_message_preview_action = getattr(
-                    value, "botMessagePreviewAction", None
-                )
-                resolved_command_id = getattr(value, "commandId", None) or getattr(
-                    value, "command_id", None
-                )
-            if bot_message_preview_action:
-                return False
-            return _match_selector(command_id, resolved_command_id)
-
-        def __call(func: Callable) -> Callable:
-            async def __handler(context: TurnContext, state: StateT) -> None:
-                action = MessagingExtensionAction.model_validate(
-                    context.activity.value or {}
-                )
-                response = await func(context, state, action)
-                if response is not None:
-                    await _send_invoke_response(context, response)
-
-            self._app.add_route(
-                __selector,
-                __handler,
-                is_invoke=True,
-                rank=rank,
-                auth_handlers=auth_handlers,
-            )
-            return func
-
-        return __call
-
-    def on_agent_message_preview_edit(
-        self,
-        command_id: CommandSelector = None,
-        *,
-        auth_handlers: Optional[list[str]] = None,
-        rank: RouteRank = RouteRank.DEFAULT,
-    ) -> Callable:
-        """Register a handler for composeExtension/submitAction with botMessagePreviewAction == 'edit'."""
-
-        def __selector(context: TurnContext) -> bool:
-            if (
-                context.activity.type != ActivityTypes.invoke
-                or context.activity.name != "composeExtension/submitAction"
-            ):
-                return False
-            value = context.activity.value or {}
-            if value.get("botMessagePreviewAction") != "edit":
-                return False
-            return _match_selector(command_id, value.get("commandId"))
-
-        def __call(func: Callable) -> Callable:
-            async def __handler(context: TurnContext, state: StateT) -> None:
-                action = MessagingExtensionAction.model_validate(
-                    context.activity.value or {}
-                )
-                response = await func(context, state, action)
-                if response is not None:
-                    await _send_invoke_response(context, response)
-
-            self._app.add_route(
-                __selector,
-                __handler,
-                is_invoke=True,
-                rank=rank,
-                auth_handlers=auth_handlers,
-            )
-            return func
-
-        return __call
-
-    def on_agent_message_preview_send(
-        self,
-        command_id: CommandSelector = None,
-        *,
-        auth_handlers: Optional[list[str]] = None,
-        rank: RouteRank = RouteRank.DEFAULT,
-    ) -> Callable:
-        """Register a handler for composeExtension/submitAction with botMessagePreviewAction == 'send'."""
-
-        def __selector(context: TurnContext) -> bool:
-            if (
-                context.activity.type != ActivityTypes.invoke
-                or context.activity.name != "composeExtension/submitAction"
-            ):
-                return False
-            value = context.activity.value or {}
-            if value.get("botMessagePreviewAction") != "send":
-                return False
-            return _match_selector(command_id, value.get("commandId"))
-
-        def __call(func: Callable) -> Callable:
-            async def __handler(context: TurnContext, state: StateT) -> None:
-                action = MessagingExtensionAction.model_validate(
-                    context.activity.value or {}
-                )
-                response = await func(context, state, action)
-                if response is not None:
-                    await _send_invoke_response(context, response)
-
-            self._app.add_route(
-                __selector,
-                __handler,
-                is_invoke=True,
-                rank=rank,
-                auth_handlers=auth_handlers,
-            )
-            return func
-
-        return __call
-
-    def on_fetch_task(
-        self,
-        command_id: CommandSelector = None,
-        *,
-        auth_handlers: Optional[list[str]] = None,
-        rank: RouteRank = RouteRank.DEFAULT,
-    ) -> Callable:
-        """Register a handler for composeExtension/fetchTask invokes."""
-
-        def __selector(context: TurnContext) -> bool:
-            return (
-                context.activity.type == ActivityTypes.invoke
-                and context.activity.name == "composeExtension/fetchTask"
-                and _match_selector(
-                    command_id,
-                    (context.activity.value or {}).get("commandId"),
-                )
-            )
-
-        def __call(func: Callable) -> Callable:
-            async def __handler(context: TurnContext, state: StateT) -> None:
-                action = MessagingExtensionAction.model_validate(
-                    context.activity.value or {}
-                )
-                response = await func(context, state, action)
-                if response is not None:
-                    await _send_invoke_response(context, response)
-
-            self._app.add_route(
-                __selector,
-                __handler,
-                is_invoke=True,
-                rank=rank,
-                auth_handlers=auth_handlers,
-            )
-            return func
-
-        return __call
-
-    def on_query_link(
-        self,
-        handler: Optional[Callable] = None,
-        *,
-        auth_handlers: Optional[list[str]] = None,
-        rank: RouteRank = RouteRank.DEFAULT,
-    ) -> Callable:
-        """Register a handler for composeExtension/queryLink invokes."""
-
-        def __selector(context: TurnContext) -> bool:
-            return (
-                context.activity.type == ActivityTypes.invoke
-                and context.activity.name == "composeExtension/queryLink"
-            )
-
-        def __register(func: Callable) -> Callable:
-            async def __handler(context: TurnContext, state: StateT) -> None:
-                query = AppBasedLinkQuery.model_validate(context.activity.value or {})
-                response = await func(context, state, query)
-                if response is not None:
-                    await _send_invoke_response(context, response)
-
-            self._app.add_route(
-                __selector,
-                __handler,
-                is_invoke=True,
-                rank=rank,
-                auth_handlers=auth_handlers,
-            )
-            return func
-
-        if handler is not None:
-            return __register(handler)
-        return __register
-
-    def on_anonymous_query_link(
-        self,
-        handler: Optional[Callable] = None,
-        *,
-        auth_handlers: Optional[list[str]] = None,
-        rank: RouteRank = RouteRank.DEFAULT,
-    ) -> Callable:
-        """Register a handler for composeExtension/anonymousQueryLink invokes."""
-
-        def __selector(context: TurnContext) -> bool:
-            return (
-                context.activity.type == ActivityTypes.invoke
-                and context.activity.name == "composeExtension/anonymousQueryLink"
-            )
-
-        def __register(func: Callable) -> Callable:
-            async def __handler(context: TurnContext, state: StateT) -> None:
-                query = AppBasedLinkQuery.model_validate(context.activity.value or {})
-                response = await func(context, state, query)
-                if response is not None:
-                    await _send_invoke_response(context, response)
-
-            self._app.add_route(
-                __selector,
-                __handler,
-                is_invoke=True,
-                rank=rank,
-                auth_handlers=auth_handlers,
-            )
-            return func
-
-        if handler is not None:
-            return __register(handler)
-        return __register
-
-    def on_query_url_setting(
-        self,
-        handler: Optional[Callable] = None,
-        *,
-        auth_handlers: Optional[list[str]] = None,
-        rank: RouteRank = RouteRank.DEFAULT,
-    ) -> Callable:
-        """Register a handler for composeExtension/querySettingUrl invokes."""
-
-        def __selector(context: TurnContext) -> bool:
-            return (
-                context.activity.type == ActivityTypes.invoke
-                and context.activity.name == "composeExtension/querySettingUrl"
-            )
-
-        def __register(func: Callable) -> Callable:
-            async def __handler(context: TurnContext, state: StateT) -> None:
-                query = MessagingExtensionQuery.model_validate(
-                    context.activity.value or {}
-                )
-                response = await func(context, state, query)
-                if response is not None:
-                    await _send_invoke_response(context, response)
-
-            self._app.add_route(
-                __selector,
-                __handler,
-                is_invoke=True,
-                rank=rank,
-                auth_handlers=auth_handlers,
-            )
-            return func
-
-        if handler is not None:
-            return __register(handler)
-        return __register
-
-    def on_configure_settings(
-        self,
-        handler: Optional[Callable] = None,
-        *,
-        auth_handlers: Optional[list[str]] = None,
-        rank: RouteRank = RouteRank.DEFAULT,
-    ) -> Callable:
-        """Register a handler for composeExtension/setting invokes."""
-
-        def __selector(context: TurnContext) -> bool:
-            return (
-                context.activity.type == ActivityTypes.invoke
-                and context.activity.name == "composeExtension/setting"
-            )
-
-        def __register(func: Callable) -> Callable:
-            async def __handler(context: TurnContext, state: StateT) -> None:
-                await func(context, state, context.activity.value)
-                await _send_invoke_response(context)
-
-            self._app.add_route(
-                __selector,
-                __handler,
-                is_invoke=True,
-                rank=rank,
-                auth_handlers=auth_handlers,
-            )
-            return func
-
-        if handler is not None:
-            return __register(handler)
-        return __register
-
-    def on_card_button_clicked(
-        self,
-        handler: Optional[Callable] = None,
-        *,
-        auth_handlers: Optional[list[str]] = None,
-        rank: RouteRank = RouteRank.DEFAULT,
-    ) -> Callable:
-        """Register a handler for composeExtension/onCardButtonClicked invokes."""
-
-        def __selector(context: TurnContext) -> bool:
-            return (
-                context.activity.type == ActivityTypes.invoke
-                and context.activity.name == "composeExtension/onCardButtonClicked"
-            )
-
-        def __register(func: Callable) -> Callable:
-            async def __handler(context: TurnContext, state: StateT) -> None:
-                await func(context, state, context.activity.value)
-                await _send_invoke_response(context)
-
-            self._app.add_route(
-                __selector,
-                __handler,
-                is_invoke=True,
-                rank=rank,
-                auth_handlers=auth_handlers,
-            )
-            return func
-
-        if handler is not None:
-            return __register(handler)
-        return __register
-
-
-class TaskModule(Generic[StateT]):
-    """
-    Route registration for Teams Task Module (task/fetch, task/submit) invoke activities.
-    Access via TeamsAgentExtension.task_module.
-    """
-
-    def __init__(self, app: AgentApplication[StateT]) -> None:
-        self._app = app
-
-    @staticmethod
-    def _get_verb(value: Optional[Any]) -> Optional[str]:
-        if not isinstance(value, dict):
-            return None
-        data = value.get("data")
-        if isinstance(data, dict):
-            return data.get("verb")
-        return None
-
-    def on_fetch(
-        self,
-        verb: CommandSelector = None,
-        *,
-        auth_handlers: Optional[list[str]] = None,
-        rank: RouteRank = RouteRank.DEFAULT,
-    ) -> Callable:
-        """Register a handler for task/fetch invokes.
-
-        :param verb: Optional verb string or regex to match against task data.
-            If None, matches all task/fetch invokes.
-        """
-
-        def __selector(context: TurnContext) -> bool:
-            if (
-                context.activity.type != ActivityTypes.invoke
-                or context.activity.name != "task/fetch"
-            ):
-                return False
-            return _match_selector(verb, TaskModule._get_verb(context.activity.value))
-
-        def __call(func: Callable) -> Callable:
-            async def __handler(context: TurnContext, state: StateT) -> None:
-                request = TaskModuleRequest.model_validate(context.activity.value or {})
-                response = await func(context, state, request)
-                if response is not None:
-                    await _send_invoke_response(context, response)
-
-            self._app.add_route(
-                __selector,
-                __handler,
-                is_invoke=True,
-                rank=rank,
-                auth_handlers=auth_handlers,
-            )
-            return func
-
-        return __call
-
-    def on_submit(
-        self,
-        verb: CommandSelector = None,
-        *,
-        auth_handlers: Optional[list[str]] = None,
-        rank: RouteRank = RouteRank.DEFAULT,
-    ) -> Callable:
-        """Register a handler for task/submit invokes.
-
-        :param verb: Optional verb string or regex to match against task data.
-            If None, matches all task/submit invokes.
-        """
-
-        def __selector(context: TurnContext) -> bool:
-            if (
-                context.activity.type != ActivityTypes.invoke
-                or context.activity.name != "task/submit"
-            ):
-                return False
-            return _match_selector(verb, TaskModule._get_verb(context.activity.value))
-
-        def __call(func: Callable) -> Callable:
-            async def __handler(context: TurnContext, state: StateT) -> None:
-                request = TaskModuleRequest.model_validate(context.activity.value or {})
-                response = await func(context, state, request)
-                if response is not None:
-                    await _send_invoke_response(context, response)
-
-            self._app.add_route(
-                __selector,
-                __handler,
-                is_invoke=True,
-                rank=rank,
-                auth_handlers=auth_handlers,
-            )
-            return func
-
-        return __call
-
-
-class Meeting(Generic[StateT]):
-    """
-    Route registration for Teams Meeting event activities.
-    Access via TeamsAgentExtension.meeting.
-    """
-
-    def __init__(self, app: AgentApplication[StateT]) -> None:
-        self._app = app
-
-    def on_start(
-        self,
-        handler: Optional[Callable] = None,
-        *,
-        auth_handlers: Optional[list[str]] = None,
-        rank: RouteRank = RouteRank.DEFAULT,
-    ) -> Callable:
-        """Register a handler for meeting start events."""
-
-        def __selector(context: TurnContext) -> bool:
-            return (
-                context.activity.type == ActivityTypes.event
-                and context.activity.name == "application/vnd.microsoft.meetingStart"
-            )
-
-        def __register(func: Callable) -> Callable:
-            async def __handler(context: TurnContext, state: StateT) -> None:
-                meeting = MeetingDetails.model_validate(context.activity.value or {})
-                await func(context, state, meeting)
-
-            self._app.add_route(
-                __selector,
-                __handler,
-                rank=rank,
-                auth_handlers=auth_handlers,
-            )
-            return func
-
-        if handler is not None:
-            return __register(handler)
-        return __register
-
-    def on_end(
-        self,
-        handler: Optional[Callable] = None,
-        *,
-        auth_handlers: Optional[list[str]] = None,
-        rank: RouteRank = RouteRank.DEFAULT,
-    ) -> Callable:
-        """Register a handler for meeting end events."""
-
-        def __selector(context: TurnContext) -> bool:
-            return (
-                context.activity.type == ActivityTypes.event
-                and context.activity.name == "application/vnd.microsoft.meetingEnd"
-            )
-
-        def __register(func: Callable) -> Callable:
-            async def __handler(context: TurnContext, state: StateT) -> None:
-                meeting = MeetingDetails.model_validate(context.activity.value or {})
-                await func(context, state, meeting)
-
-            self._app.add_route(
-                __selector,
-                __handler,
-                rank=rank,
-                auth_handlers=auth_handlers,
-            )
-            return func
-
-        if handler is not None:
-            return __register(handler)
-        return __register
-
-    def on_participants_join(
-        self,
-        handler: Optional[Callable] = None,
-        *,
-        auth_handlers: Optional[list[str]] = None,
-        rank: RouteRank = RouteRank.DEFAULT,
-    ) -> Callable:
-        """Register a handler for meeting participant join events."""
-
-        def __selector(context: TurnContext) -> bool:
-            return (
-                context.activity.type == ActivityTypes.event
-                and context.activity.name
-                == "application/vnd.microsoft.meetingParticipantJoin"
-            )
-
-        def __register(func: Callable) -> Callable:
-            async def __handler(context: TurnContext, state: StateT) -> None:
-                details = MeetingParticipantsEventDetails.model_validate(
-                    context.activity.value or {}
-                )
-                await func(context, state, details)
-
-            self._app.add_route(
-                __selector,
-                __handler,
-                rank=rank,
-                auth_handlers=auth_handlers,
-            )
-            return func
-
-        if handler is not None:
-            return __register(handler)
-        return __register
-
-    def on_participants_leave(
-        self,
-        handler: Optional[Callable] = None,
-        *,
-        auth_handlers: Optional[list[str]] = None,
-        rank: RouteRank = RouteRank.DEFAULT,
-    ) -> Callable:
-        """Register a handler for meeting participant leave events."""
-
-        def __selector(context: TurnContext) -> bool:
-            return (
-                context.activity.type == ActivityTypes.event
-                and context.activity.name
-                == "application/vnd.microsoft.meetingParticipantLeave"
-            )
-
-        def __register(func: Callable) -> Callable:
-            async def __handler(context: TurnContext, state: StateT) -> None:
-                details = MeetingParticipantsEventDetails.model_validate(
-                    context.activity.value or {}
-                )
-                await func(context, state, details)
-
-            self._app.add_route(
-                __selector,
-                __handler,
-                rank=rank,
-                auth_handlers=auth_handlers,
-            )
-            return func
-
-        if handler is not None:
-            return __register(handler)
-        return __register
-
 
 class TeamsAgentExtension(Generic[StateT]):
     """
